feat(tests): add comprehensive unit tests for AML v2 rules
- Introduce new unit tests for AML v2, verifying multiple rule outcomes
- Implement helper function for fetching user details in test scenarios
- Refactor existing company sanctions test cases to utilize helper function
