+++ services/workflows-service/src/rule-engine/core/test/idv-check.unit.test.ts
@@ -8,6 +8,10 @@ import {
 } from '@ballerine/common';
 import { createRuleEngine } from '../rule-engine';
 
+const helpers = {
+  getEndUserById: async (id: string) => {},
+};
+
 describe('IDV_CHECK operator', () => {
   it('should pass when decision status is declined', async () => {
     const mockData = {
@@ -40,7 +44,7 @@ describe('IDV_CHECK operator', () => {
       ],
     };
 
-    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData);
+    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData, helpers);
 
     expect(validationResults).toBeDefined();
     expect(validationResults).toHaveLength(1);
@@ -78,7 +82,7 @@ describe('IDV_CHECK operator', () => {
       ],
     };
 
-    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData);
+    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData, helpers);
 
     expect(validationResults).toBeDefined();
     expect(validationResults).toHaveLength(1);
@@ -116,7 +120,7 @@ describe('IDV_CHECK operator', () => {
       ],
     };
 
-    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData);
+    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData, helpers);
 
     expect(validationResults).toBeDefined();
     expect(validationResults).toHaveLength(1);
@@ -164,7 +168,7 @@ describe('IDV_CHECK operator', () => {
       ],
     };
 
-    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData);
+    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData, helpers);
 
     expect(validationResults).toBeDefined();
     expect(validationResults).toHaveLength(1);
@@ -202,7 +206,7 @@ describe('IDV_CHECK operator', () => {
       ],
     };
 
-    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData);
+    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData, helpers);
 
     expect(validationResults).toBeDefined();
     expect(validationResults).toHaveLength(1);
@@ -239,7 +243,7 @@ describe('IDV_CHECK operator', () => {
       ],
     };
 
-    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData);
+    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData, helpers);
 
     expect(validationResults).toBeDefined();
     expect(validationResults).toHaveLength(1);
@@ -266,7 +270,7 @@ describe('IDV_CHECK operator', () => {
       ],
     };
 
-    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData);
+    const validationResults: RuleResultSet = await createRuleEngine(ruleSet).run(mockData, helpers);
 
     expect(validationResults).toBeDefined();
     expect(validationResults).toHaveLength(1);
