+++ services/workflows-service/src/rule-engine/rule-engine.service.intg.test.ts
@@ -1,6 +1,6 @@
+import { RuleSet } from '@ballerine/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { RuleEngineService } from './rule-engine.service';
-import { RuleSet } from '@ballerine/common';
 
 describe('RuleEngineService', () => {
   let service: RuleEngineService;
@@ -12,6 +12,10 @@ describe('RuleEngineService', () => {
 
     service = module.get<RuleEngineService>(RuleEngineService);
   });
+
+  const helpers = {
+    getEndUserById: async (id: string) => {},
+  };
 
   it('should run the IN_CASE_INSENSITIVE rule successfully', async () => {
     const rules: RuleSet = {
@@ -37,7 +41,7 @@ describe('RuleEngineService', () => {
       array: ['THIS IS AN OWNERSHIP COMPANY'],
     };
 
-    const result = await service.run(rules, formData);
+    const result = await service.run(rules, formData, helpers);
 
     result.forEach(r => {
       expect(r.status).toEqual('PASSED');
